fixing a bit more the rf eu conversion

+++ src/main/java/de/keridos/floodlights/tileentity/TileEntityElectricFloodlight.java
@@ -54,8 +54,8 @@ public class TileEntityElectricFloodlight extends TileEntityFLElectric implement
             if (inventory[0] != null) {
                 if (ModCompatibility.IC2Loaded) {
                     if (inventory[0].getItem() instanceof IElectricItem) {
-                        double dischargeValue = (storage.getMaxEnergyStored() - storage.getEnergyStored()) / 8;
-                        storage.modifyEnergyStored(MathUtil.truncateDoubleToInt(8 * ElectricItem.manager.discharge(inventory[0], dischargeValue, 4, false, true, false)));
+                        double dischargeValue = (storage.getMaxEnergyStored() - (double) storage.getEnergyStored()) / 8.0D;
+                        storage.modifyEnergyStored(MathUtil.truncateDoubleToInt(8.0D * ElectricItem.manager.discharge(inventory[0], dischargeValue, 4, false, true, false)));
                     }
                 }
                 if (inventory[0].getItem() instanceof IEnergyContainerItem) {
@@ -75,8 +75,8 @@ public class TileEntityElectricFloodlight extends TileEntityFLElectric implement
                         world.setBlockMetadataWithNotify(this.xCoord, this.yCoord, this.zCoord, world.getBlockMetadata(this.xCoord, this.yCoord, this.zCoord) + 6, 2);
                     }
                 }
-                if (storageEU >= realEnergyUsage / 8) {
-                    storageEU -= realEnergyUsage / 8;
+                if (storageEU >= (double) realEnergyUsage / 8.0D) {
+                    storageEU -= (double) realEnergyUsage / 8.0D;
                 } else {
                     storage.modifyEnergyStored(-realEnergyUsage);
                 }
